第三版:序列化器与SPI机制

+++ example-common/src/main/java/com/zzy/example/common/service/UserService.java
@@ -13,7 +13,8 @@ public interface UserService {
     /**
      * 新方法 - 获取数字
      */
-    default short getNumber(){
+    default Integer getNumber(){
         return 1;
     }
+
 }
